Move formatSassError to dedicated module
